Rewrote the example and added a translation
Using for as an example is terible! It's the only name that isn't changed.
I fully understad that its probalby suposed to be `för` but whit a `o` as a `ö`

I also added an keword translated version of the example.

PS. I fully understad the irony of writing the commit message in english ;)